Add MemberRepository
