+++ lib/page/night/night_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:wordwolf/provider/presentation_providers.dart';
-import 'package:wordwolf/repository/room_repository.dart';
+import 'package:wordwolf/repository/member_repository.dart';
 
 class NightPage extends ConsumerWidget {
   const NightPage({Key? key, required this.roomId}) : super(key: key);
@@ -16,7 +16,7 @@ class NightPage extends ConsumerWidget {
       body: Center(
         child: members.when(
           data: (mem) {
-            final livingMem = ref.read(roomRepositoryProvider).getLivingMembers(mem);
+            final livingMem = ref.read(memberRepositoryProvider).getLivingMembers(mem);
             if (livingMem[livingMem.indexWhere((e) => e.userId == 'gpt')].isLive ==
                 false) {
               return const Text('村人');
